refactor(menu): decouple MenuBlock from its content via MenuItemData list
- Extract MenuItemData to menu_item_data.dart (public, reusable data class)
- Extract MenuBlock + _MenuTile to menu_block.dart; widget now accepts
  List<MenuItemData> and has no knowledge of what the items do
- FullScreenMenuPage builds the item list at the call site (inside a Builder
  for colorScheme access) and passes it to MenuBlock — no more hardcoded
  callbacks baked into the block widget itself
