+++ lib/features/menu/full_screen_menu_page.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:trainlog_app/features/menu/menu_block.dart';
 import 'package:trainlog_app/features/menu/menu_explore_card.dart';
+import 'package:trainlog_app/features/menu/menu_item_data.dart';
 import 'package:trainlog_app/features/menu/menu_summary_card.dart';
 import 'package:trainlog_app/app/app_colors.dart';
 import 'package:trainlog_app/l10n/app_localizations.dart';
@@ -55,22 +57,46 @@ class FullScreenMenuPage extends StatelessWidget {
                     const SizedBox(height: 24),
                     _SectionHeader(label: loc.menuMenuSectionTitle),
                     const SizedBox(height: 8),
-                    _MenuBlock(
-                      loc: loc,
-                      onInboxTap: onInboxTap,
-                      onTrainlogStatusTap: onTrainlogStatusTap,
-                      onAboutTap: () => onPageTap(AppPageId.about),
-                      onLogout: () async {
-                        final settings = context.read<SettingsProvider>();
-                        final trips = context.read<TripsProvider>();
-                        final scaffMsg = ScaffoldMessenger.of(context);
-                        onClose();
-                        await context.read<TrainlogProvider>().logout(settings, trips);
-                        scaffMsg.showSnackBar(
-                          SnackBar(content: Text(loc.loggedOut)),
-                        );
-                      },
-                    ),
+                    Builder(builder: (context) {
+                      final cs = Theme.of(context).colorScheme;
+                      return MenuBlock(items: [
+                        MenuItemData(
+                          icon: AdaptiveIcons.inbox,
+                          iconBg: AppColors.amber,
+                          label: loc.menuInboxTitle,
+                          onTap: onInboxTap,
+                        ),
+                        MenuItemData(
+                          icon: AdaptiveIcons.ok,
+                          iconBg: AppColors.blue,
+                          label: loc.trainglogStatusPageTitle,
+                          onTap: onTrainlogStatusTap,
+                        ),
+                        MenuItemData(
+                          icon: AdaptiveIcons.info,
+                          iconBg: AppColors.navy,
+                          label: loc.menuAboutTitle,
+                          onTap: () => onPageTap(AppPageId.about),
+                        ),
+                        MenuItemData(
+                          icon: AdaptiveIcons.logout,
+                          iconBg: cs.error,
+                          label: loc.logoutButton,
+                          labelColor: cs.error,
+                          isDestructive: true,
+                          onTap: () async {
+                            final settings = context.read<SettingsProvider>();
+                            final trips = context.read<TripsProvider>();
+                            final scaffMsg = ScaffoldMessenger.of(context);
+                            onClose();
+                            await context.read<TrainlogProvider>().logout(settings, trips);
+                            scaffMsg.showSnackBar(
+                              SnackBar(content: Text(loc.loggedOut)),
+                            );
+                          },
+                        ),
+                      ]);
+                    }),
                     const SizedBox(height: 32),
                   ],
                 ),
@@ -220,158 +246,6 @@ class _ExploreGrid extends StatelessWidget {
       children: items
           .map((d) => ExploreCard(data: d, onTap: () => onPageTap(d.id)))
           .toList(),
-    );
-  }
-}
-
-// ── Menu block (single grouped list) ─────────────────────────────────────────
-
-class _MenuBlock extends StatelessWidget {
-  final AppLocalizations loc;
-  final VoidCallback onInboxTap;
-  final VoidCallback onTrainlogStatusTap;
-  final VoidCallback onAboutTap;
-  final VoidCallback onLogout;
-
-  const _MenuBlock({
-    required this.loc,
-    required this.onInboxTap,
-    required this.onTrainlogStatusTap,
-    required this.onAboutTap,
-    required this.onLogout,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final cs = Theme.of(context).colorScheme;
-
-    final items = <_MenuItemData>[
-      _MenuItemData(
-        icon: AdaptiveIcons.inbox,
-        iconBg: AppColors.amber,
-        label: loc.menuInboxTitle,
-        onTap: onInboxTap,
-      ),
-      _MenuItemData(
-        icon: AdaptiveIcons.ok,
-        iconBg: AppColors.blue,
-        label: loc.trainglogStatusPageTitle,
-        onTap: onTrainlogStatusTap,
-      ),
-      _MenuItemData(
-        icon: AdaptiveIcons.info,
-        iconBg: Colors.grey,
-        label: loc.menuAboutTitle,
-        onTap: onAboutTap,
-      ),
-      _MenuItemData(
-        icon: AdaptiveIcons.logout,
-        iconBg: cs.error,
-        label: loc.logoutButton,
-        labelColor: cs.error,
-        onTap: onLogout,
-        isDestructive: true,
-      ),
-    ];
-
-    return Container(
-      decoration: BoxDecoration(
-        color: cs.surface,
-        borderRadius: BorderRadius.circular(14),
-        border: Border.all(color: cs.outlineVariant, width: 1),
-      ),
-      child: Column(
-        children: [
-          for (int i = 0; i < items.length; i++) ...[
-            _MenuTile(
-              data: items[i],
-              isFirst: i == 0,
-              isLast: i == items.length - 1,
-            ),
-            if (i < items.length - 1)
-              Divider(
-                height: 1,
-                thickness: 1,
-                indent: 16 + 28 + 12, // align with label start
-                endIndent: 0,
-                color: cs.outlineVariant,
-              ),
-          ],
-        ],
-      ),
-    );
-  }
-}
-
-class _MenuItemData {
-  final IconData icon;
-  final Color iconBg;
-  final String label;
-  final Color? labelColor;
-  final VoidCallback onTap;
-  final bool isDestructive;
-
-  const _MenuItemData({
-    required this.icon,
-    required this.iconBg,
-    required this.label,
-    required this.onTap,
-    this.labelColor,
-    this.isDestructive = false,
-  });
-}
-
-class _MenuTile extends StatelessWidget {
-  final _MenuItemData data;
-  final bool isFirst;
-  final bool isLast;
-
-  const _MenuTile({
-    required this.data,
-    required this.isFirst,
-    required this.isLast,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final cs = Theme.of(context).colorScheme;
-    final resolvedTextColor = data.labelColor ?? cs.onSurface;
-
-    return InkWell(
-      onTap: data.onTap,
-      borderRadius: BorderRadius.vertical(
-        top: isFirst ? const Radius.circular(14) : Radius.zero,
-        bottom: isLast ? const Radius.circular(14) : Radius.zero,
-      ),
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-        child: Row(
-          children: [
-            Container(
-              width: 28,
-              height: 28,
-              decoration: BoxDecoration(
-                color: data.iconBg,
-                borderRadius: BorderRadius.circular(6),
-              ),
-              child: Icon(data.icon, color: Colors.white, size: 15),
-            ),
-            const SizedBox(width: 12),
-            Expanded(
-              child: Text(
-                data.label,
-                style: TextStyle(
-                  fontSize: 15,
-                  fontWeight: FontWeight.w500,
-                  color: resolvedTextColor,
-                ),
-              ),
-            ),
-            if (!data.isDestructive)
-              Icon(Icons.chevron_right, size: 20, color: cs.onSurfaceVariant),
-          ],
-        ),
-      ),
     );
   }
 }
